--- project3mongodbmongoose/models/index.js
@@ -1,5 +1,0 @@
-module.exports={
-    Student: require("./student.js"),
-    Attendance: require("./attendance.js"),
-    Class: require('./class')
-}
