--- project3mongodbmongoose/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "project3mongodbmongoose",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "admin-bro": "^1.4.2",
-    "admin-bro-expressjs": "^0.3.0",
-    "admin-bro-mongoose": "^0.3.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "mongoose": "^5.7.12"
-  }
-}
