--- project3mongodbmongoose/routes/index.js
@@ -1,8 +1,0 @@
-console.log('routes connected');
-const router = require("express").Router()
-
-router.get("/", (req,res)=>{
-    res.send("the html is connected")
-})
-
-module.exports=router;
