+++ routes/index.js
@@ -1,13 +1,8 @@
-const path = require("path");
-const router = require("express").Router();
-const apiRoutes = require("./api");
+console.log('routes connected');
+const router = require("express").Router()
 
-// API Routes
-router.use("/api", apiRoutes);
+router.get("/", (req,res)=>{
+    res.send("the html is connected")
+})
 
-// If no API routes are hit, send the React app
-router.use(function(req, res) {
-  res.sendFile(path.join(__dirname, "../client/build/index.html"));
-});
-
-module.exports = router;
+module.exports=router;
